렌더링 파이프라인 과정 구조화(Mesh, Shader, InputLayout, ConstantBuffer)

+++ Engine_Source/yaTime.cpp
@@ -24,7 +24,7 @@ namespace ya
 	{
 		QueryPerformanceCounter(&mCurFrequency);
 
-		double differnceFrequency = mCurFrequency.QuadPart - mPrevFrequency.QuadPart;
+		double differnceFrequency = (double)(mCurFrequency.QuadPart - mPrevFrequency.QuadPart);
 
 		mDeltaTime = differnceFrequency / mCpuFrequency.QuadPart;
